# [improve][java-client] Replace ScheduledExecutor to improve performance of message consumption